Улучшение и рефакторинг PerformanceTimer
Перемещен класс PerformanceTimer в новое пространство имен
`ReportEngine.Shared.Utils` для улучшения структуры проекта.

Добавлена проверка на `null` для параметров методов `Measure`
и `MeasureAsync`, чтобы предотвратить исключения
`NullReferenceException`.

Логика измерения времени выполнения обернута в блок
`try-finally` для гарантии остановки таймера даже при
возникновении исключений.

Удален вывод в консоль (`Console.WriteLine`), оставлен только
вывод в отладочный поток (`Debug.WriteLine`).

В методе `MeasureAsync` добавлен вызов `ConfigureAwait(false)`
для предотвращения захвата контекста синхронизации.
